requirements.txt
openai
langchain
langchain_openai
huggingface_hub
python-dotenv
streamlit